Extract `google-charts` service
