+++ app/components/google-jsapi.js
@@ -1,46 +1,12 @@
 import Component from '@ember/component';
-
-function createEvent(name) {
-  let event = document.createEvent('Event');
-  event.initEvent(name, true, true);
-  return event;
-}
+import { inject as service } from '@ember/service';
 
 export default Component.extend({
+  googleCharts: service(),
+
   tagName: '',
 
   didInsertElement() {
-    loadJsApi()
-      .then(api => loadCoreChart(api))
-      .then(() => {
-        window.googleChartsLoaded = true;
-        document.dispatchEvent(createEvent('googleChartsLoaded'));
-      });
+    this.googleCharts.load();
   },
 });
-
-async function loadScript(src) {
-  await new Promise((resolve, reject) => {
-    const script = document.createElement('script');
-    script.src = src;
-    script.onload = resolve;
-    script.onerror = reject;
-    document.body.appendChild(script);
-  });
-}
-
-async function loadJsApi() {
-  if (!window.google) {
-    await loadScript('https://www.google.com/jsapi');
-  }
-  return window.google;
-}
-
-async function loadCoreChart(api) {
-  await new Promise(resolve => {
-    api.load('visualization', '1.0', {
-      packages: ['corechart'],
-      callback: resolve,
-    });
-  });
-}
